Changed JSON output on list view to a table. Added controls to set new coords and radius incl. reloading of POI list.
