+++ angular-maps-frontend/src/app/service/poi-format.service.ts
@@ -1,0 +1,53 @@
+import { Injectable } from '@angular/core';
+import { Sanitizer } from '../util/sanitization.util';
+
+@Injectable({ providedIn: 'root' })
+export class PoiFormatService {
+
+    private iconPhone = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-telephone" viewBox="0 0 16 16">
+  <path d="M3.654 1.328a.678.678 0 0 0-1.015-.063L1.605 2.3c-.483.484-.661 1.169-.45 1.77a17.6 17.6 0 0 0 4.168 6.608 17.6 17.6 0 0 0 6.608 4.168c.601.211 1.286.033 1.77-.45l1.034-1.034a.678.678 0 0 0-.063-1.015l-2.307-1.794a.68.68 0 0 0-.58-.122l-2.19.547a1.75 1.75 0 0 1-1.657-.459L5.482 8.062a1.75 1.75 0 0 1-.46-1.657l.548-2.19a.68.68 0 0 0-.122-.58zM1.884.511a1.745 1.745 0 0 1 2.612.163L6.29 2.98c.329.423.445.974.315 1.494l-.547 2.19a.68.68 0 0 0 .178.643l2.457 2.457a.68.68 0 0 0 .644.178l2.189-.547a1.75 1.75 0 0 1 1.494.315l2.306 1.794c.829.645.905 1.87.163 2.611l-1.034 1.034c-.74.74-1.846 1.065-2.877.702a18.6 18.6 0 0 1-7.01-4.42 18.6 18.6 0 0 1-4.42-7.009c-.362-1.03-.037-2.137.703-2.877z"/>
+</svg>`;
+
+    constructor(private sanitizer: Sanitizer) { }
+
+    formatDetails(details: string | undefined): string {
+        if (!details) return '';
+
+        const detailChunks = details.split(/, |\n/).map(s => (s || '').trim()).filter(s => s.length > 0);
+
+        const formatted = detailChunks.map(chunk => {
+            if (this.sanitizer.isSafeUrl(chunk)) {
+                return this.formatForLink(chunk);
+            }
+            return this.formatForPhone(chunk);
+        });
+
+        return formatted.join('<br>');
+    }
+
+    formatForLink(text: string): string {
+        const t = (text || '').trim();
+        if (this.sanitizer.isSafeUrl(t)) {
+            const href = t.toLowerCase().startsWith('http') ? t : `https://${t}`;
+            const sanitizedHref = this.sanitizer.sanitizeText(href, this.sanitizer.maxHref);
+            return `<a href="${sanitizedHref}" target="_blank" rel="noopener">${sanitizedHref}</a>`;
+        }
+        return this.sanitizer.sanitizeText(t, this.sanitizer.maxText);
+    }
+
+    formatForPhone(text: string): string {
+        const raw = (text || '').trim();
+        if (!raw) return '';
+
+        if (raw.startsWith('+49')) {
+            return this.iconPhone + ' ' + this.sanitizer.sanitizeText(raw, this.sanitizer.maxPhone);
+        }
+
+        if (raw.startsWith('Tel.:')) {
+            const num = raw.replace('Tel.:', '').trim();
+            return this.iconPhone + ' ' + this.sanitizer.sanitizeText(num, this.sanitizer.maxPhone);
+        }
+
+        return this.sanitizer.sanitizeText(raw, this.sanitizer.maxText);
+    }
+}
